requirements.txt
python-telegram-bot[job-queue]==20.7
groq==0.9.0
aiohttp==3.9.5
beautifulsoup4==4.12.3
lxml==5.2.2
aiosqlite==0.20.0
apscheduler==3.10.4
python-dotenv==1.0.1
fake-useragent==1.5.1
stripe==9.12.0
fastapi==0.111.0
uvicorn==0.30.1
pytz==2024.1
